refactor:SurveyUser 컴포넌트 단계별 분리
