+++ src/pages/Survey/SurveyUser.jsx
@@ -1,15 +1,19 @@
 import { useState } from 'react';
-import { Box, Typography, Container } from '@mui/material';
-import { SurveyHeader } from '@/components/Common/SurveyHeader/SurveyHeader';
+import { Container, Box } from '@mui/material';
 import { useNavigate } from 'react-router-dom';
-import InputText from '@/components/Common/InputText/InputText';
-import NumberPicker from '@/components/Common/NumberPicker/NumberPicker';
-import RadioButton from '@/components/Common/RadioButton/RadioButton';
+import { SurveyHeader } from '@/components/Common/SurveyHeader/SurveyHeader';
 import Button from '@/components/Common/Button/Button';
 import { Modal } from '@/components/Common/Modal/Modal';
+import Step1 from '@/components/Survey/UserSteps/Step1';
+import Step2 from '@/components/Survey/UserSteps/Step2';
+import Step3 from '@/components/Survey/UserSteps/Step3';
+import Step4 from '@/components/Survey/UserSteps/Step4';
+import Step5 from '@/components/Survey/UserSteps/Step5';
+import Step6 from '@/components/Survey/UserSteps/Step6';
+import Step7 from '@/components/Survey/UserSteps/Step7';
 import uploadProfileButton from '/images/upload-dog-button.png';
 
-function SurveyUser() {
+const SurveyUser = () => {
   const navigate = useNavigate();
   const [step, setStep] = useState(1);
   const [petInfo, setPetInfo] = useState({
@@ -64,7 +68,7 @@ function SurveyUser() {
   };
 
   const handleSaveProfile = () => {
-    // 여기서 프로필 저장 로직 수행
+    // 프로필 저장 로직
     return true; // 저장 성공 시 true 반환
   };
 
@@ -92,12 +96,6 @@ function SurveyUser() {
     }
   };
 
-  const handleGoHome = () => {
-    if (handleSaveProfile()) {
-      navigate('/home');
-    }
-  };
-
   const handleNextStep = () => {
     setStep((prev) => prev + 1);
   };
@@ -109,241 +107,9 @@ function SurveyUser() {
       navigate('/survey');
     }
   };
-
-  const renderStep = () => {
-    switch (step) {
-      case 1:
-        return (
-          <Box sx={{ mt: 8 }}>
-            <Typography variant="h5" sx={{ fontWeight: 'bold', mb: 4 }}>
-              반려견의 이름을 입력해주세요.
-            </Typography>
-            <InputText
-              size="large"
-              placeholder="이름을 입력해주세요"
-              value={petInfo.name}
-              onChange={(e) =>
-                setPetInfo((prev) => ({ ...prev, name: e.target.value }))
-              }
-            />
-          </Box>
-        );
-
-      case 2:
-        return (
-          <Box sx={{ mt: 8 }}>
-            <Typography variant="h5" sx={{ fontWeight: 'bold', mb: 4 }}>
-              반려견의 나이를 입력해주세요.
-            </Typography>
-            <Box
-              sx={{
-                display: 'flex',
-                gap: 2,
-                alignItems: 'center',
-                justifyContent: 'center',
-              }}
-            >
-              <NumberPicker
-                value={petInfo.age}
-                onChange={(value) =>
-                  setPetInfo((prev) => ({ ...prev, age: value }))
-                }
-                label="년"
-                max={20}
-              />
-              <NumberPicker
-                value={petInfo.ageMonth}
-                onChange={(value) =>
-                  setPetInfo((prev) => ({ ...prev, ageMonth: value }))
-                }
-                label="개월"
-                max={11}
-              />
-            </Box>
-          </Box>
-        );
-
-      case 3:
-        return (
-          <Box sx={{ mt: 8 }}>
-            <Typography variant="h5" sx={{ fontWeight: 'bold', mb: 4 }}>
-              반려견의 몸무게를 선택해주세요.
-            </Typography>
-            <Box
-              sx={{
-                display: 'flex',
-                justifyContent: 'center',
-              }}
-            >
-              <NumberPicker
-                value={petInfo.weight}
-                onChange={(value) =>
-                  setPetInfo((prev) => ({ ...prev, weight: value }))
-                }
-                label="kg"
-                max={100}
-              />
-            </Box>
-          </Box>
-        );
-
-      case 4:
-        return (
-          <Box sx={{ mt: 8 }}>
-            <Typography variant="h5" sx={{ fontWeight: 'bold', mb: 4 }}>
-              반려견의 견종을 입력해주세요.
-            </Typography>
-            <InputText
-              size="large"
-              placeholder="견종을 입력해주세요"
-              value={petInfo.breed}
-              onChange={(e) =>
-                setPetInfo((prev) => ({ ...prev, breed: e.target.value }))
-              }
-            />
-          </Box>
-        );
-
-      case 5:
-        return (
-          <Box sx={{ mt: 8 }}>
-            <Typography variant="h5" sx={{ fontWeight: 'bold', mb: 4 }}>
-              반려견은 어떤 성별인가요?
-            </Typography>
-            <Typography sx={{ mb: 2 }}>성별</Typography>
-            <Box sx={{ display: 'flex', gap: '12px', mb: 4 }}>
-              <RadioButton
-                label="남아"
-                selected={petInfo.gender === '남아'}
-                onChange={() =>
-                  setPetInfo((prev) => ({ ...prev, gender: '남아' }))
-                }
-                size="large"
-              />
-              <RadioButton
-                label="여아"
-                selected={petInfo.gender === '여아'}
-                onChange={() =>
-                  setPetInfo((prev) => ({ ...prev, gender: '여아' }))
-                }
-                size="large"
-              />
-            </Box>
-            <Typography sx={{ mb: 2 }}>중성화 여부</Typography>{' '}
-            <Box sx={{ display: 'flex', gap: '12px' }}>
-              <RadioButton
-                label="중성화 했어요."
-                selected={petInfo.neutered}
-                onChange={() =>
-                  setPetInfo((prev) => ({ ...prev, neutered: true }))
-                }
-                size="large"
-              />
-            </Box>
-          </Box>
-        );
-      case 6:
-        return (
-          <Box sx={{ mt: 8 }}>
-            <Typography variant="h5" sx={{ fontWeight: 'bold', mb: 4 }}>
-              반려견은 어떤 특징을 가졌나요?
-            </Typography>
-            <Box display="flex" flexDirection="column" gap={3}>
-              {' '}
-              {Object.entries(petInfo.characteristics).map(
-                ([trait, checked]) => (
-                  <Box key={trait}>
-                    <RadioButton
-                      label={trait}
-                      selected={checked}
-                      onChange={() => {
-                        if (trait === '없음') {
-                          const newCharacteristics = Object.keys(
-                            petInfo.characteristics
-                          ).reduce(
-                            (acc, key) => ({
-                              ...acc,
-                              [key]: key === '없음',
-                            }),
-                            {}
-                          );
-                          setPetInfo((prev) => ({
-                            ...prev,
-                            characteristics: newCharacteristics,
-                            otherCharacteristic: '',
-                          }));
-                        } else {
-                          setPetInfo((prev) => ({
-                            ...prev,
-                            characteristics: {
-                              ...prev.characteristics,
-                              [trait]: !checked,
-                              없음: false,
-                            },
-                          }));
-                        }
-                      }}
-                      size="large"
-                    />
-                    {trait === '기타' && checked && (
-                      <Box sx={{ mt: 2, mb: 1 }}>
-                        {' '}
-                        <InputText
-                          size="large"
-                          placeholder="기타 특징을 적어주세요. (최대30자)"
-                          value={petInfo.otherCharacteristic}
-                          onChange={(e) =>
-                            setPetInfo((prev) => ({
-                              ...prev,
-                              otherCharacteristic: e.target.value,
-                            }))
-                          }
-                        />
-                      </Box>
-                    )}
-                  </Box>
-                )
-              )}
-            </Box>
-          </Box>
-        );
-      case 7:
-        return (
-          <Box
-            sx={{
-              mt: 8,
-              display: 'flex',
-              flexDirection: 'column',
-              alignItems: 'center',
-            }}
-          >
-            <Typography variant="h5" sx={{ fontWeight: 'bold', mb: 4 }}>
-              프로필 사진을 등록해주세요.
-            </Typography>
-            <Box
-              sx={{
-                width: '200px',
-                height: '200px',
-                display: 'flex',
-                justifyContent: 'center',
-                alignItems: 'center',
-              }}
-            >
-              <img
-                src={uploadProfileButton}
-                alt="프로필 업로드"
-                style={{
-                  width: '100%',
-                  height: '100%',
-                  objectFit: 'contain',
-                }}
-              />
-            </Box>
-          </Box>
-        );
-      default:
-        return null;
-    }
+  const handleGoHome = () => {
+    handleSaveProfile();
+    navigate('/home');
   };
 
   return (
@@ -356,7 +122,19 @@ function SurveyUser() {
       />
 
       <Container maxWidth="sm" sx={{ px: 2, pb: 8 }}>
-        {renderStep()}
+        {step === 1 && <Step1 petInfo={petInfo} setPetInfo={setPetInfo} />}
+        {step === 2 && <Step2 petInfo={petInfo} setPetInfo={setPetInfo} />}
+        {step === 3 && <Step3 petInfo={petInfo} setPetInfo={setPetInfo} />}
+        {step === 4 && <Step4 petInfo={petInfo} setPetInfo={setPetInfo} />}
+        {step === 5 && <Step5 petInfo={petInfo} setPetInfo={setPetInfo} />}
+        {step === 6 && <Step6 petInfo={petInfo} setPetInfo={setPetInfo} />}
+        {step === 7 && (
+          <Step7
+            petInfo={petInfo}
+            setPetInfo={setPetInfo}
+            uploadProfileButton={uploadProfileButton}
+          />
+        )}
 
         <Box
           sx={{
@@ -399,6 +177,6 @@ function SurveyUser() {
       </Container>
     </>
   );
-}
+};
 
 export default SurveyUser;
